First hack at full signed offset mode handling for long distances out of range.

+++ fpc_hosted/aarch64/hdrc.pas
@@ -41,7 +41,7 @@ const
 
   bitsperbyte = 8; {packing factor}
   byte = 1; 
-  short = 2;
+  half = 2;
   word = 4;
   long = 8;
   quad = 16;
